Move files related to the club
