+++ app/src/main/java/com/swmaestro/roundup/com/swmaestro/roundup/club/Feed.java
@@ -1,5 +1,5 @@
 
-package com.swmaestro.roundup;
+package com.swmaestro.roundup.com.swmaestro.roundup.club;
 
 public class Feed {
