+++ app/src/main/java/com/swmaestro/roundup/com/swmaestro/roundup/club/RecyclerViewAdapter.java
@@ -1,5 +1,5 @@
 
-package com.swmaestro.roundup;
+package com.swmaestro.roundup.com.swmaestro.roundup.club;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
@@ -10,6 +10,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
+import com.swmaestro.roundup.R;
 
 import java.util.ArrayList;
 import java.util.List;
